requirements/base.txt
Django==4.1
python-dotenv==0.20.0
djangorestframework==3.13.1